backend/requirements.txt
aiohttp-retry==2.9.1
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.3.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.115.12
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
openai>=1.0
packaging==25.0
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
validators==0.34.0
watchfiles==1.0.5
weaviate-client==4.14.4
websockets==15.0.1
